Add files via upload
Please create a git repository and push the code there and send across the link when complete.
The task is to create a mock Netflix like page with editable rows of titles using any framework you want (preferably React).
The top row will be the ‘My List’ row and will display the titles included in ‘mylist’ array. The second row will be the recommendation row which will display the titles included in ‘recommendations’ array.
Hovering over an image in the ‘My List’ row should create a “Remove” button displayed below the image. Clicking on the button should remove the title from the “My List” row.
Hovering over an image in the recommendations row should create an “Add” button displayed below the image. Clicking on the button should add the title to the “My List” row.
The list of titles in the “My List” row should be displayed at the bottom of the page.